Split Rlimiter into a standalone module to reduce the amount of stuff it pulls in and bring back the fd closing code.
